Rename package from "framework" to "correctness" and fix some variable and build target names.
